+++ src/main/java/Controller/ProductController.java
@@ -1,14 +1,7 @@
 package Controller;
 
-import DTO.ProductDTO;
-import DTO.TotebagDTO;
-import Entity.Product;
-import Service.impl.ProductService;
+import Service.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
